Refactor of the example mod
