+++ gradle.properties
@@ -1,16 +1,16 @@
 # Done to increase the memory available to gradle.
-org.gradle.jvmargs=-Xmx1G
+org.gradle.jvmargs=-Xmx2G
 
 # Fabric Properties
 	# check these on https://fabricmc.net/develop
 	minecraft_version=1.19.1
-	yarn_mappings=1.19.1+build.1
+	yarn_mappings=1.19.1+build.5
 	loader_version=0.14.8
 
-# Mod Properties
-	mod_version = 1.0.0
-	maven_group = com.example
-	archives_base_name = fabric-example-mod
+	#Fabric api
+	fabric_version=0.58.5+1.19.1
 
-# Dependencies
-	fabric_version=0.58.4+1.19.1
+# Mod Properties
+	mod_version = 0.0.1
+	maven_group = xyz.srgnis
+	archives_base_name = bodyhealthsystem
