+++ src/main/java/xyz/srgnis/bodyhealthsystem/BHSMain.java
@@ -1,14 +1,14 @@
-package net.fabricmc.example;
+package xyz.srgnis.bodyhealthsystem;
 
 import net.fabricmc.api.ModInitializer;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-public class ExampleMod implements ModInitializer {
+public class BHSMain implements ModInitializer {
 	// This logger is used to write text to the console and the log file.
 	// It is considered best practice to use your mod id as the logger's name.
 	// That way, it's clear which mod wrote info, warnings, and errors.
-	public static final Logger LOGGER = LoggerFactory.getLogger("modid");
+	public static final Logger LOGGER = LoggerFactory.getLogger("bodyhealthsystem");
 
 	@Override
 	public void onInitialize() {
