+++ src/main/resources/bodyhealthsystem.mixins.json
@@ -1,12 +1,9 @@
 {
   "required": true,
   "minVersion": "0.8",
-  "package": "net.fabricmc.example.mixin",
+  "package": "xyz.srgnis.bodyhealthsystem.mixin",
   "compatibilityLevel": "JAVA_17",
   "mixins": [
-  ],
-  "client": [
-    "ExampleMixin"
   ],
   "injectors": {
     "defaultRequire": 1
